server/requirements.txt
Django==4.2.4
django-cors-headers==4.2.0
eventlet==0.33.3
python-socketio==5.8.0